requirements.txt
pylrc
pymediainfo
ffmpeg-python
yt-dlp
